backend/requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.6.1
Django==2.2
django-cors-headers==2.5.2
djangorestframework==3.9.2
Markdown==3.1
pycparser==2.19
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
